pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for ORON watch marketplace"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]
